Floating Subnet Config: Wildcards and Defaults
Allow to configure a default subnet for floating pools configuration in the CloudProfile.
This will be used to assign the router (if not an existing one passed) to the matching subnet.
Furthermore enable the selection of floating subnet by wildcards

This change require a Terraformer with openstack provider v1.37 or higher.
The floating subnet id is removed from the Infrastrucutre state as it was not used.