Improve LoadBalancerClass validation
Allow to change the name and purpose of load balancer classes in `.controlPlaneConfig.loadBalancerClasses[]`. The load balancer classes configuration need still to be semantically equal with the load balancer classes from the CloudProfile.

Add validation to check if there are no load balancer classes with same name and if there are multiple default or private loadbalancer classes.