Enable propagating pod routes to nodes without overlay network.
In case only the virtual network is used for an overlay-free pod network,
all packets may go over the virtual router. This may introduce additional
latency and may overload the virtual router.
It is possible, though, to propagate the pod routes to all nodes so that
a direct route is used instead.