Fix indentation in CCM ServiceMonitor (#789)
The `honorLabels`, `authorization`, `scheme`, and `tlsConfig` fields are
part of the Endpoint API, not the Spec. This indentation error caused the
prometheus-operator to generate a scrape job configuration that specified
http instead of https (because the scheme field was "missing". This in turn
caused CCM-down alerts to fire because the scrape failed with a 400 error.

Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>