Adapt the retry logic for the resolv.conf copy script to deal with (#1161)
ancient systemd releases.

Some dated linux distributions use old systemd releases, which do not
allow units to specify RestartForceExitStatus and Type=oneshot to be set
at the same time. To allow the systemd unit to be executed also in those
systems we use the less elegant solution using a loop.